NeighborlyAPI/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
pymongo